producer/requirements.txt
confluent-kafka==2.3.0
kafka-python==2.0.2
requests==2.31.0
python-dotenv==1.0.0
minio==7.1.15
cassandra-driver==3.28.0